refactor(booking-fragment): split submit-login into hook
