+++ fragments/booking-fragment/src/initial/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './submit-form.hook.js'
