+++ fragments/booking-fragment/src/initial/hooks/submit-form.interface.ts
@@ -1,0 +1,12 @@
+type SubmitFormHookProps = {
+  name: string
+  phone: string
+  selectedCarBody: string
+  selectedRadius: string
+  selectedRepairTypes: Array<string>
+  comment: string
+}
+
+type SubmitFormHookType = (props: SubmitFormHookProps) => Promise<void>
+
+export type { SubmitFormHookType }
